Added Folder-per-Component Structure
