+++ src/components/LanguageSwitcher/LanguageSwitcher.css
@@ -1,0 +1,4 @@
+.language-switcher {
+  min-width: 120px;
+  margin-bottom: 16px;
+}
