+++ src/components/WelcomeScreen/WelcomeScreen.css
@@ -1,0 +1,25 @@
+.welcome-paper {
+  padding: 2rem;
+  margin-top: 2rem;
+  max-width: 400px;
+  margin-left: auto;
+  margin-right: auto;
+  background: #1a1a1a;
+}
+
+.welcome-motion {
+  opacity: 0;
+  transform: translateY(40px);
+}
+
+.welcome-motion-animate {
+  opacity: 1 !important;
+  transform: translateY(0) !important;
+  transition: opacity 0.5s, transform 0.5s;
+}
+
+.welcome-title {
+  color: #d32f2f;
+  font-weight: 700;
+  margin-bottom: 1rem;
+}
